PREBOOT: romset into own file
